Created a well organised folder structure to maintain
