+++ frontend/app/(tabs)/(home)/filterTabs/places/MapScreen.tsx
@@ -2,8 +2,8 @@ import React, { useState } from "react";
 import { View, StyleSheet, TouchableOpacity } from "react-native";
 import Map from "./Map"; // Adjust path based on your folder structure
 import { useBottomTabBarHeight } from "@react-navigation/bottom-tabs";
-import { fetchFoodPlaces } from "@/app/api/fetchFoodPlaces";
-import { fetchEntertainmentPlaces } from "@/app/api/fetchEntertainmentPlaces";
+import { fetchFoodPlaces } from "@/services/api/fetchFoodPlaces";
+import { fetchEntertainmentPlaces } from "@/services/api/fetchEntertainmentPlaces";
 
 import { MaterialCommunityIcons, MaterialIcons } from "@expo/vector-icons";
