+++ frontend/app/_layout.tsx
@@ -1,12 +1,11 @@
-import { Stack } from 'expo-router';
+import { Stack } from "expo-router";
 
 export default function RootLayout() {
   return (
     <Stack screenOptions={{ headerShown: false }}>
       <Stack.Screen name="(tabs)" />
       <Stack.Screen name="(auth)" />
-      <Stack.Screen name="(profile)" />
+      {/* <Stack.Screen name="(profile)" /> */}
     </Stack>
-    
   );
 }
