+++ frontend/services/api/fetchEntertainmentPlaces.js
