+++ frontend/services/api/fetchUserDetails.ts
