+++ frontend/tsconfig.json
@@ -12,6 +12,7 @@
     ".expo/types/**/*.ts",
     "expo-env.d.ts",
     "app/(tabs)/(home)/filterTabs/FilterTabBar.jsx",
-    "app/(tabs)/(home)/inputSection/InputSection.jsx"
-, "app/(tabs)/(profile)/index.jsx"  ]
+    "app/(tabs)/(home)/inputSection/InputSection.jsx",
+    "app/(tabs)/(profile)/index.jsx"
+  ]
 }
